[API] Doctrine filters moved below filters namespace
