+++ src/Sylius/Bundle/ApiBundle/Filter/Doctrine/TaxonFilter.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\ApiBundle\Filter;
+namespace Sylius\Bundle\ApiBundle\Filter\Doctrine;
 
 use ApiPlatform\Core\Api\IriConverterInterface;
 use ApiPlatform\Core\Bridge\Doctrine\Orm\Filter\AbstractContextAwareFilter;
