+++ src/Sylius/Bundle/ApiBundle/Resources/config/api_resources/ExchangeRate.xml
@@ -24,7 +24,7 @@
             <collectionOperation name="admin_get">
                 <attribute name="method">GET</attribute>
                 <attribute name="filters">
-                    <attribute>Sylius\Bundle\ApiBundle\Doctrine\Filter\ExchangeRateFilter</attribute>
+                    <attribute>Sylius\Bundle\ApiBundle\Filter\Doctrine\ExchangeRateFilter</attribute>
                 </attribute>
                 <attribute name="normalization_context">
                     <attribute name="groups">admin:exchange_rate:read</attribute>
