+++ src/Sylius/Bundle/ApiBundle/Resources/config/api_resources/Product.xml
@@ -30,7 +30,7 @@
                     <attribute>sylius.api.product_name_filter</attribute>
                     <attribute>sylius.api.product_order_filter</attribute>
                     <attribute>sylius.api.product_taxon_code_filter</attribute>
-                    <attribute>Sylius\Bundle\ApiBundle\Doctrine\Filter\TranslationOrderNameAndLocaleFilter</attribute>
+                    <attribute>Sylius\Bundle\ApiBundle\Filter\Doctrine\TranslationOrderNameAndLocaleFilter</attribute>
                 </attribute>
                 <attribute name="normalization_context">
                     <attribute name="groups">admin:product:read</attribute>
@@ -44,9 +44,9 @@
                     <attribute>sylius.api.product_name_filter</attribute>
                     <attribute>sylius.api.product_order_filter</attribute>
                     <attribute>sylius.api.product_taxon_code_filter</attribute>
-                    <attribute>Sylius\Bundle\ApiBundle\Doctrine\Filter\ProductPriceOrderFilter</attribute>
-                    <attribute>Sylius\Bundle\ApiBundle\Doctrine\Filter\TranslationOrderNameAndLocaleFilter</attribute>
-                    <attribute>Sylius\Bundle\ApiBundle\Filter\TaxonFilter</attribute>
+                    <attribute>Sylius\Bundle\ApiBundle\Filter\Doctrine\ProductPriceOrderFilter</attribute>
+                    <attribute>Sylius\Bundle\ApiBundle\Filter\Doctrine\TranslationOrderNameAndLocaleFilter</attribute>
+                    <attribute>Sylius\Bundle\ApiBundle\Filter\Doctrine\TaxonFilter</attribute>
                 </attribute>
                 <attribute name="normalization_context">
                     <attribute name="groups">shop:product:read</attribute>
