+++ src/Sylius/Bundle/ApiBundle/Resources/config/api_resources/ProductVariant.xml
@@ -72,7 +72,7 @@
                 <attribute name="path">/shop/product-variants</attribute>
                 <attribute name="filters">
                     <attribute>sylius.api.product_variant_product_filter</attribute>
-                    <attribute>Sylius\Bundle\ApiBundle\Doctrine\Filter\ProductVariantOptionValueFilter</attribute>
+                    <attribute>Sylius\Bundle\ApiBundle\Filter\Doctrine\ProductVariantOptionValueFilter</attribute>
                 </attribute>
                 <attribute name="normalization_context">
                     <attribute name="groups">shop:product_variant:read</attribute>
